Autoprovision users in all SAML-enabled subsites
In a multisite setup, the autoprovisioning mechanism only provisioned the user
into the site they are actively logging into (current site). However, they then
exist network-wide (they're $existing_user), which prevented the user from ever
autoprovisioning into other sites.
One option would have been to also autoprovision - into the current site -
existing users that are not yet member of it. but that solution would require
the user to explicitly login at each subsite separately, which is inconvenient
and confusing for the user: a user logged into one wordpress site, is also
logged into all the other sites in the network, even when not a member.
Hence it is easier to autoprovision users directly into all currently existing
subsites that have the WP-SAML plugin enabled. By checking this each time the
user is logging in, they also get autoprovisioned when a new site is created.